header, cookis, cors done

+++ src/db/db_user.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm.session import Session
 from request_models.user_request import UserBase
 from db.models import DbUser
 from db.hash import Hash
+from fastapi import HTTPException, status
 
 def create_user(db: Session, request: UserBase):
     new_user = DbUser(
@@ -18,10 +19,15 @@ def get_all_uesrs(db: Session):
     return db.query(DbUser).all()
 
 def get_user_by_id(db: Session, id: int):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User with id {id} not found")
+    return user
 
 def update_user(db:Session, id:int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User with id {id} not found")
     user.update({
 
         DbUser.username : request.username,
@@ -34,6 +40,8 @@ def update_user(db:Session, id:int, request: UserBase):
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="User with id {id} not found")
     db.delete(user)
     db.commit()
     return "Ok"
